REQ-1: Implement UI for last night's sleep data
In this commit, I implemented the UI for displaying last night's sleep data. If the data is available, it is displayed; if not, an "Add Last Night's Sleep Data" button is shown. Clicking this button opens a modal window with a form.
Additionally, I removed the reducer and service for logging in, as I decided to add the token to the env file for security purposes. Moreover, user authentication is not a requirement for this test task.